# AI-Doc-Manager/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dms-backend",
    "description": "## Document Management System (DMS) Backend\n\nREST API for uploading, versioning, reviewing, and approving documents.  Includes RAG-powered semantic search and Q&A, AI-driven runbook generation, and role-based access control.\n\n### Authentication\n\nMost endpoints require a JWT **Bearer** token obtained from `POST /api/v1/auth/login`.  Include the token in the `Authorization` header:\n\n    Authorization: Bearer <your_token>\n\n### Error Responses\n\nAll error responses follow the same structure:\n\n```json\n{\"detail\": \"Human-readable message\", \"code\": \"machine_readable_code\", \"request_id\": \"uuid\"}\n```\n\n### Pagination\n\nList endpoints return paginated results with `page`, `page_size`, and `total` fields.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/admin/users": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Create a new user (admin)",
        "description": "Creates a new user with the specified roles.  Requires admin-level permissions.  The username must be unique.",
        "operationId": "create_user_api_v1_admin_users_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource conflict (e.g. duplicate username or invalid state transition)"
          },
          "400": {
            "description": "Bad request — malformed input or missing required fields"
          },
          "401": {
            "description": "Missing or invalid authentication credentials"
          },
          "403": {
            "description": "Authenticated user lacks the required permission"
          },
          "404": {
            "description": "The requested resource was not found"
          },
          "422": {
            "description": "Request body or parameters failed validation"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all users (admin)",
        "description": "Returns a paginated list of all registered users.  Requires admin-level permissions.",
        "operationId": "list_users_route_api_v1_admin_users_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-based)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-based)"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Items per page (1–200)",
              "default": 50,
              "title": "Page Size"
            },
            "description": "Items per page (1–200)"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of user records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request — malformed input or missing required fields"
          },
          "401": {
            "description": "Missing or invalid authentication credentials"
          },
          "403": {
            "description": "Authenticated user lacks the required permission"
          },
          "422": {
            "description": "Request body or parameters failed validation"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get a user by ID (admin)",
        "description": "Returns the details of a specific user including their assigned roles.  Requires admin-level permissions.",
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The requested user's details and roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "400": {
            "description": "Bad request — malformed input or missing required fields"
          },
          "401": {
            "description": "Missing or invalid authentication credentials"
          },
          "403": {
            "description": "Authenticated user lacks the required permission"
          },
          "422": {
            "description": "Request body or parameters failed validation"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/roles": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Assign a role to a user (admin)",
        "description": "Grants an additional role to a user.  The role must already exist in the system.  Requires admin-level permissions.",
        "operationId": "assign_role_api_v1_admin_users__user_id__roles_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The user's updated profile with all assigned roles",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found"
          },
          "409": {
            "description": "Resource conflict (e.g. duplicate username or invalid state transition)"
          },
          "400": {
            "description": "Bad request — malformed input or missing required fields"
          },
          "401": {
            "description": "Missing or invalid authentication credentials"
          },
          "403": {
            "description": "Authenticated user lacks the required permission"
          },
          "422": {
            "description": "Request body or parameters failed validation"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/roles/{role_name}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Remove a role from a user (admin)",
        "description": "Revokes a specific role from a user.  Requires admin-level permissions.",
        "operationId": "remove_role_api_v1_admin_users__user_id__roles__role_name__delete",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "role_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The user's updated profile after role removal",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource was not found"
          },
          "401": {
            "description": "Missing or invalid authentication credentials"
          },
          "403": {
            "description": "Authenticated user lacks the required permission"
          },
          "500": {
            "description": "Internal server error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all roles (admin)",
        "description": "Returns all defined security roles and their descriptions.  Requires admin-level permissions.",
        "operationId": "list_roles_route_api_v1_admin_roles_get",
        "responses": {
          "200": {
            "description": "List of all role definitions",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleResponse"
                  },
                  "type": "array",
                  "title": "Response List Roles Route Api V1 Admin Roles Get"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignRoleRequest": {
        "properties": {
          "role_name": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Role Name",
            "description": "Name of the role to assign (e.g. 'editor', 'admin')",
            "examples": [
              "editor"
            ]
          }
        },
        "type": "object",
        "required": [
          "role_name"
        ],
        "title": "AssignRoleRequest",
        "description": "Payload to assign an additional role to a user.",
        "example": {
          "role_name": "editor"
        }
      },
...
      "CreateUserRequest": {
        "properties": {
          "username": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Username",
            "description": "Desired username (3–100 characters, unique across the system)",
            "examples": [
              "jdoe"
            ]
          },
          "password": {
            "type": "string",
            "maxLength": 255,
            "minLength": 8,
            "title": "Password",
            "description": "Account password (minimum 8 characters)",
            "examples": [
              "secure-initial-password"
            ]
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Role names to assign at creation time (e.g. ['editor', 'viewer'])",
            "examples": [
              [
                "editor",
                "viewer"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "CreateUserRequest",
        "description": "Payload for an admin to create a new user with initial roles.",
        "example": {
          "password": "secure-initial-password",
          "roles": [
            "editor",
            "viewer"
          ],
          "username": "jdoe"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoleResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique role identifier (UUID v4)",
            "examples": [
              "660e8400-e29b-41d4-a716-446655440001"
            ]
          },
          "role_name": {
            "type": "string",
            "title": "Role Name",
            "description": "Unique role name (e.g. 'admin', 'editor', 'viewer')",
            "examples": [
              "editor"
            ]
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Human-readable description of what this role grants access to",
            "examples": [
              "Can view, edit, and submit documents for review"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "role_name"
        ],
        "title": "RoleResponse",
        "description": "A security role definition.",
        "example": {
          "description": "Can view, edit, and submit documents for review",
          "id": "660e8400-e29b-41d4-a716-446655440001",
          "role_name": "editor"
        }
      },
...
      "UserListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserResponse"
            },
            "type": "array",
            "title": "Items",
            "description": "User records for the current page"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of users matching the query (across all pages)",
            "examples": [
              42
            ]
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Current page number (1-based)",
            "examples": [
              1
            ]
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size",
            "description": "Number of items per page",
            "examples": [
              50
            ]
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "UserListResponse",
        "description": "Paginated list of users."
      },
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Unique user identifier (UUID v4)",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "username": {
            "type": "string",
            "title": "Username",
            "description": "Registered username",
            "examples": [
              "jdoe"
            ]
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Roles",
            "description": "Role names currently assigned to this user",
            "examples": [
              [
                "editor",
                "viewer"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "roles"
        ],
        "title": "UserResponse",
        "description": "Summary of a user account — returned by admin endpoints.",
        "example": {
          "id": "550e8400-e29b-41d4-a716-446655440000",
          "roles": [
            "editor",
            "viewer"
          ],
          "username": "jdoe"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from POST /api/v1/auth/login.  Prefix the token value with nothing — just the raw JWT string."
      }
    }
  },
...
  "security": [
    {
      "BearerAuth": []
    }
  ]
}